Add PointCollectionDescriptor
This is a descriptor for a "mesh" that is a collection of points.
Remapping can only be performed with this "mesh" as the destination,
since the "grid cells" around each point have vertices that coincide
with the cell center (thus zero area).

Point collections like this are useful, for example, for remapping
3D data sets to transects.